refactor(rviz_panel): Split into smaller compilation units
This should greatly improve recompilation speed and reduce overall
memory use (especially when limiting the number of build cores)
