+++ .pre-commit-config.yaml
@@ -1,13 +1,13 @@
 ci:
+  autoupdate_schedule: quarterly
   autofix_prs: false
-
 repos:
 - repo: meta
   hooks:
   - id: check-useless-excludes
   - id: check-hooks-apply
 - repo: https://github.com/pre-commit/mirrors-clang-format
-  rev: v20.1.4
+  rev: v20.1.5
   hooks:
   - id: clang-format
     'types_or': [c++, c]
@@ -28,3 +28,9 @@ repos:
   rev: v0.6.13
   hooks:
   - id: cmake-format
+- repo: https://github.com/compilerla/conventional-pre-commit
+  rev: v4.2.0
+  hooks:
+    - id: conventional-pre-commit
+      stages: [commit-msg]
+      args: [--verbose] # optional: list of Conventional Commits types to allow e.g. [feat, fix, ci, chore, test]
