+++ mc_rtc_rviz_panel/src/Panel.cpp
@@ -3,32 +3,9 @@
  */
 
 #include "Panel.h"
+#include "PanelImpl.h"
 
-#include "ArrayInputWidget.h"
-#include "ArrayLabelWidget.h"
-#include "ArrowInteractiveMarkerWidget.h"
-#include "ButtonWidget.h"
-#include "CheckboxWidget.h"
-#include "ComboInputWidget.h"
 #include "ConnectionDialog.h"
-#include "DisplayTrajectoryWidget.h"
-#include "ForceInteractiveMarkerWidget.h"
-#include "FormElement.h"
-#include "FormWidget.h"
-#include "GenericInputWidget.h"
-#include "InteractiveMarkerWidget.h"
-#include "LabelWidget.h"
-#include "NumberSliderWidget.h"
-#include "PlotTabWidget.h"
-#include "Point3DInteractiveMarkerWidget.h"
-#include "PolygonMarkerWidget.h"
-#include "PolyhedronMarkerWidget.h"
-#include "RobotModelDisplay.h"
-#include "SchemaWidget.h"
-#include "TableWidget.h"
-#include "TransformInteractiveMarkerWidget.h"
-#include "VisualWidget.h"
-#include "XYThetaInteractiveMarkerWidget.h"
 
 #include <boost/filesystem.hpp>
 namespace bfs = boost::filesystem;
@@ -110,30 +87,6 @@ double getTimeout()
 }
 
 } // namespace
-
-struct PanelImpl
-{
-  PanelImpl()
-  {
-#ifdef MC_RTC_ROS_IS_ROS2
-    nh_ = rclcpp::Node::make_shared("mc_rtc_rviz_panel");
-    marker_array_pub_ = nh_->create_publisher<MarkerArray>("/mc_rtc_rviz", 0);
-    int_server_ = std::make_shared<InteractiveMarkerServer>("mc_rtc_rviz_interactive_markers", nh_);
-#else
-    int_server_ = std::make_shared<InteractiveMarkerServer>("mc_rtc_rviz_interactive_markers");
-    marker_array_pub_ = nh_.advertise<MarkerArray>("/mc_rtc_rviz", 0);
-#endif
-  }
-#ifdef MC_RTC_ROS_IS_ROS2
-  mc_rtc::NodeHandlePtr nh_;
-  rclcpp::Publisher<MarkerArray>::SharedPtr marker_array_pub_;
-#else
-  ros::NodeHandle nh_;
-  ros::Publisher marker_array_pub_;
-#endif
-  std::shared_ptr<InteractiveMarkerServer> int_server_;
-  MarkerArray marker_array_;
-};
 
 Panel::Panel(QWidget * parent)
 : CategoryWidget(ClientWidgetParam{*this, parent, {{}, "ROOT"}}),
@@ -303,373 +256,6 @@ void Panel::category(const std::vector<std::string> & parent, const std::string 
   Q_EMIT signal_category(parent, category);
 }
 
-void Panel::label(const WidgetId & id, const std::string & dataIn)
-{
-  Q_EMIT signal_label(id, dataIn);
-}
-
-void Panel::array_label(const WidgetId & id, const std::vector<std::string> & labels, const Eigen::VectorXd & dataIn)
-{
-  Q_EMIT signal_array_label(id, labels, dataIn);
-}
-
-void Panel::button(const WidgetId & id)
-{
-  Q_EMIT signal_button(id);
-}
-
-void Panel::checkbox(const WidgetId & id, bool state)
-{
-  Q_EMIT signal_checkbox(id, state);
-}
-
-void Panel::string_input(const WidgetId & id, const std::string & dataIn)
-{
-  Q_EMIT signal_string_input(id, dataIn);
-}
-
-void Panel::integer_input(const WidgetId & id, int dataIn)
-{
-  Q_EMIT signal_integer_input(id, dataIn);
-}
-
-void Panel::number_input(const WidgetId & id, double dataIn)
-{
-  Q_EMIT signal_number_input(id, dataIn);
-}
-
-void Panel::number_slider(const WidgetId & id, double dataIn, double min, double max)
-{
-  Q_EMIT signal_number_slider(id, dataIn, min, max);
-}
-
-void Panel::array_input(const WidgetId & id, const std::vector<std::string> & inputs, const Eigen::VectorXd & dataIn)
-{
-  Q_EMIT signal_array_input(id, inputs, dataIn);
-}
-
-void Panel::combo_input(const WidgetId & id, const std::vector<std::string> & values, const std::string & dataIn)
-{
-  Q_EMIT signal_combo_input(id, values, dataIn);
-}
-
-void Panel::data_combo_input(const WidgetId & id, const std::vector<std::string> & ref, const std::string & dataIn)
-{
-  Q_EMIT signal_data_combo_input(id, ref, dataIn);
-}
-
-void Panel::point3d(const WidgetId & id,
-                    const WidgetId & requestId,
-                    bool ro,
-                    const Eigen::Vector3d & pos,
-                    const mc_rtc::gui::PointConfig & config)
-{
-  Q_EMIT signal_point3d(id, requestId, ro, pos, config);
-}
-
-void Panel::trajectory(const WidgetId & id,
-                       const std::vector<Eigen::Vector3d> & points,
-                       const mc_rtc::gui::LineConfig & config)
-{
-  Q_EMIT signal_trajectory(id, points, config);
-}
-
-void Panel::trajectory(const WidgetId & id,
-                       const std::vector<sva::PTransformd> & points,
-                       const mc_rtc::gui::LineConfig & config)
-{
-  Q_EMIT signal_trajectory(id, points, config);
-}
-
-void Panel::trajectory(const WidgetId & id, const Eigen::Vector3d & point, const mc_rtc::gui::LineConfig & config)
-{
-  Q_EMIT signal_trajectory(id, point, config);
-}
-
-void Panel::trajectory(const WidgetId & id, const sva::PTransformd & point, const mc_rtc::gui::LineConfig & config)
-{
-  Q_EMIT signal_trajectory(id, point, config);
-}
-
-void Panel::polygon(const WidgetId & id,
-                    const std::vector<std::vector<Eigen::Vector3d>> & polygons,
-                    const mc_rtc::gui::Color & color)
-{
-  polygon(id, polygons, mc_rtc::gui::LineConfig{color});
-}
-
-void Panel::polygon(const WidgetId & id,
-                    const std::vector<std::vector<Eigen::Vector3d>> & polygons,
-                    const mc_rtc::gui::LineConfig & config)
-{
-  Q_EMIT signal_polygon(id, polygons, config);
-}
-
-void Panel::polyhedron(const WidgetId & id,
-                       const std::vector<std::array<Eigen::Vector3d, 3>> & triangles,
-                       const std::vector<std::array<mc_rtc::gui::Color, 3>> & colors,
-                       const mc_rtc::gui::PolyhedronConfig & c)
-{
-  Q_EMIT signal_polyhedron(id, triangles, colors, c);
-}
-
-void Panel::force(const WidgetId & id,
-                  const WidgetId & requestId,
-                  const sva::ForceVecd & force_,
-                  const sva::PTransformd & start,
-                  const mc_rtc::gui::ForceConfig & forceConfig,
-                  bool ro)
-{
-  Q_EMIT signal_force(id, requestId, force_, start, forceConfig, ro);
-}
-
-void Panel::arrow(const WidgetId & id,
-                  const WidgetId & requestId,
-                  const Eigen::Vector3d & start,
-                  const Eigen::Vector3d & end,
-                  const mc_rtc::gui::ArrowConfig & config,
-                  bool ro)
-{
-  Q_EMIT signal_arrow(id, requestId, start, end, config, ro);
-}
-
-void Panel::rotation(const WidgetId & id, const WidgetId & requestId, bool ro, const sva::PTransformd & pos)
-{
-  Q_EMIT signal_rotation(id, requestId, ro, pos);
-}
-
-void Panel::transform(const WidgetId & id, const WidgetId & requestId, bool ro, const sva::PTransformd & pos)
-{
-  Q_EMIT signal_transform(id, requestId, ro, pos);
-}
-
-void Panel::xytheta(const WidgetId & id,
-                    const WidgetId & requestId,
-                    bool ro,
-                    const Eigen::Vector3d & vec,
-                    double altitude)
-{
-  Q_EMIT signal_xytheta(id, requestId, ro, vec, altitude);
-}
-
-void Panel::schema(const WidgetId & id, const std::string & schema)
-{
-  Q_EMIT signal_schema(id, schema);
-}
-
-void Panel::table_start(const WidgetId & id, const std::vector<std::string> & header)
-{
-  Q_EMIT signal_table_start(id, header);
-}
-
-void Panel::table_row(const WidgetId & id, const std::vector<std::string> & dataIn)
-{
-  Q_EMIT signal_table_row(id, dataIn);
-}
-
-void Panel::table_end(const WidgetId & id)
-{
-  Q_EMIT signal_table_end(id);
-}
-
-void Panel::robot(const WidgetId & id,
-                  const std::vector<std::string> & parameters,
-                  const std::vector<std::vector<double>> & q,
-                  const sva::PTransformd & posW)
-{
-  Q_EMIT signal_robot(id, parameters, q, posW);
-}
-
-void Panel::visual(const WidgetId & id, const rbd::parsers::Visual & visual, const sva::PTransformd & pos)
-{
-  Q_EMIT signal_visual(id, visual, pos);
-}
-
-void Panel::form(const WidgetId & id)
-{
-  Q_EMIT signal_form(id);
-}
-
-void Panel::form_checkbox(const WidgetId & formId, const std::string & name, bool required, bool def, bool def_from_user)
-{
-  Q_EMIT signal_form_checkbox(formId, name, required, def, def_from_user);
-}
-
-void Panel::form_integer_input(const WidgetId & formId,
-                               const std::string & name,
-                               bool required,
-                               int def,
-                               bool def_from_user)
-{
-  Q_EMIT signal_form_integer_input(formId, name, required, def, def_from_user);
-}
-
-void Panel::form_number_input(const WidgetId & formId,
-                              const std::string & name,
-                              bool required,
-                              double def,
-                              bool def_from_user)
-{
-  Q_EMIT signal_form_number_input(formId, name, required, def, def_from_user);
-}
-
-void Panel::form_string_input(const WidgetId & formId,
-                              const std::string & name,
-                              bool required,
-                              const std::string & def,
-                              bool def_from_user)
-{
-  Q_EMIT signal_form_string_input(formId, name, required, def, def_from_user);
-}
-
-void Panel::form_array_input(const WidgetId & formId,
-                             const std::string & name,
-                             bool required,
-                             const Eigen::VectorXd & def,
-                             bool fixed_size,
-                             bool def_from_user)
-{
-  Q_EMIT signal_form_array_input(formId, name, required, def, fixed_size, def_from_user);
-}
-
-void Panel::form_combo_input(const WidgetId & formId,
-                             const std::string & name,
-                             bool required,
-                             const std::vector<std::string> & values,
-                             bool send_index,
-                             int def)
-{
-  Q_EMIT signal_form_combo_input(formId, name, required, values, send_index, def);
-}
-
-void Panel::form_data_combo_input(const WidgetId & formId,
-                                  const std::string & name,
-                                  bool required,
-                                  const std::vector<std::string> & ref,
-                                  bool send_index)
-{
-  Q_EMIT signal_form_data_combo_input(formId, name, required, ref, send_index);
-}
-
-void Panel::form_point3d_input(const WidgetId & formId,
-                               const std::string & name,
-                               bool required,
-                               const Eigen::Vector3d & default_,
-                               bool default_from_user,
-                               bool interactive)
-{
-  Q_EMIT signal_form_point3d_input(formId, name, required, default_, default_from_user, interactive);
-}
-
-void Panel::form_rotation_input(const WidgetId & formId,
-                                const std::string & name,
-                                bool required,
-                                const sva::PTransformd & default_,
-                                bool default_from_user,
-                                bool interactive)
-{
-  Q_EMIT signal_form_rotation_input(formId, name, required, default_, default_from_user, interactive);
-}
-
-void Panel::form_transform_input(const WidgetId & formId,
-                                 const std::string & name,
-                                 bool required,
-                                 const sva::PTransformd & default_,
-                                 bool default_from_user,
-                                 bool interactive)
-{
-  Q_EMIT signal_form_transform_input(formId, name, required, default_, default_from_user, interactive);
-}
-
-void Panel::start_form_object_input(const std::string & name, bool required)
-{
-  Q_EMIT signal_start_form_object_input(name, required);
-}
-
-void Panel::end_form_object_input()
-{
-  Q_EMIT signal_end_form_object_input();
-}
-
-void Panel::start_form_generic_array_input(const std::string & name,
-                                           bool required,
-                                           std::optional<std::vector<mc_rtc::Configuration>> data)
-{
-  Q_EMIT signal_start_form_generic_array_input(name, required, data);
-}
-
-void Panel::end_form_generic_array_input()
-{
-  Q_EMIT signal_end_form_generic_array_input();
-}
-
-void Panel::start_form_one_of_input(const std::string & name,
-                                    bool required,
-                                    const std::optional<std::pair<size_t, mc_rtc::Configuration>> & data)
-{
-  Q_EMIT signal_start_form_one_of_input(name, required, data);
-}
-
-void Panel::end_form_one_of_input()
-{
-  Q_EMIT signal_end_form_one_of_input();
-}
-
-void Panel::start_plot(uint64_t id, const std::string & title)
-{
-  Q_EMIT signal_start_plot(id, title);
-}
-
-void Panel::plot_setup_xaxis(uint64_t id, const std::string & legend, const mc_rtc::gui::plot::Range & range)
-{
-  Q_EMIT signal_plot_setup_xaxis(id, legend, range);
-}
-
-void Panel::plot_setup_yaxis_left(uint64_t id, const std::string & legend, const mc_rtc::gui::plot::Range & range)
-{
-  Q_EMIT signal_plot_setup_yaxis_left(id, legend, range);
-}
-
-void Panel::plot_setup_yaxis_right(uint64_t id, const std::string & legend, const mc_rtc::gui::plot::Range & range)
-{
-  Q_EMIT signal_plot_setup_yaxis_right(id, legend, range);
-}
-
-void Panel::plot_point(uint64_t id,
-                       uint64_t did,
-                       const std::string & legend,
-                       double x,
-                       double y,
-                       mc_rtc::gui::Color color,
-                       mc_rtc::gui::plot::Style style,
-                       mc_rtc::gui::plot::Side side)
-{
-  Q_EMIT signal_plot_point(id, did, legend, x, y, color, style, side);
-}
-
-void Panel::plot_polygon(uint64_t id,
-                         uint64_t did,
-                         const std::string & legend,
-                         const mc_rtc::gui::plot::PolygonDescription & polygon,
-                         mc_rtc::gui::plot::Side side)
-{
-  Q_EMIT signal_plot_polygon(id, did, legend, polygon, side);
-}
-
-void Panel::plot_polygons(uint64_t id,
-                          uint64_t did,
-                          const std::string & legend,
-                          const std::vector<mc_rtc::gui::plot::PolygonDescription> & polygons,
-                          mc_rtc::gui::plot::Side side)
-{
-  Q_EMIT signal_plot_polygons(id, did, legend, polygons, side);
-}
-
-void Panel::end_plot(uint64_t id)
-{
-  Q_EMIT signal_end_plot(id);
-}
-
 void Panel::got_category(const std::vector<std::string> & parent, const std::string & category)
 {
   auto & tree = get_category(parent);
@@ -680,432 +266,6 @@ void Panel::got_category(const std::vector<std::string> & parent, const std::str
     tree.sub_trees_[category].parent = cat;
   }
   tree.sub_trees_[category].parent->seen();
-}
-
-void Panel::got_label(const WidgetId & id, const std::string & dataIn)
-{
-  auto & w = get_widget<LabelWidget>(id);
-  w.update(dataIn);
-}
-
-void Panel::got_array_label(const WidgetId & id,
-                            const std::vector<std::string> & labels,
-                            const Eigen::VectorXd & dataIn)
-{
-  auto & w = get_widget<ArrayLabelWidget>(id, labels);
-  w.update(dataIn);
-}
-
-void Panel::got_button(const WidgetId & id)
-{
-  get_widget<ButtonWidget>(id);
-}
-
-void Panel::got_checkbox(const WidgetId & id, bool state)
-{
-  auto & w = get_widget<CheckboxWidget>(id);
-  w.update(state);
-}
-
-void Panel::got_string_input(const WidgetId & id, const std::string & dataIn)
-{
-  auto & w = get_widget<StringInputWidget>(id);
-  w.update(dataIn);
-}
-
-void Panel::got_integer_input(const WidgetId & id, int dataIn)
-{
-  auto & w = get_widget<IntegerInputWidget>(id);
-  w.update(dataIn);
-}
-
-void Panel::got_number_input(const WidgetId & id, double dataIn)
-{
-  auto & w = get_widget<NumberInputWidget>(id);
-  w.update(dataIn);
-}
-
-void Panel::got_number_slider(const WidgetId & id, double dataIn, double min, double max)
-{
-  auto & w = get_widget<NumberSliderWidget>(id, min, max);
-  w.update(dataIn, min, max);
-}
-
-void Panel::got_array_input(const WidgetId & id,
-                            const std::vector<std::string> & inputs,
-                            const Eigen::VectorXd & dataIn)
-{
-  auto & w = get_widget<ArrayInputWidget>(id, inputs);
-  w.update(dataIn);
-}
-
-void Panel::got_combo_input(const WidgetId & id, const std::vector<std::string> & values, const std::string & dataIn)
-{
-  auto & w = get_widget<ComboInputWidget>(id, values);
-  w.update(dataIn, values);
-}
-
-void Panel::got_data_combo_input(const WidgetId & id,
-                                 const std::vector<std::string> & values,
-                                 const std::string & dataIn)
-{
-  auto & w = get_widget<ComboInputWidget>(id, data_, values);
-  w.update(dataIn, data_, values);
-}
-
-void Panel::got_point3d(const WidgetId & id,
-                        const WidgetId & requestId,
-                        bool ro,
-                        const Eigen::Vector3d & pos,
-                        const mc_rtc::gui::PointConfig & config)
-{
-  auto label = latestWidget_;
-  auto & w = get_widget<Point3DInteractiveMarkerWidget>(id, requestId, impl_->int_server_, config, !ro, label);
-  w.update(pos);
-}
-
-void Panel::got_rotation(const WidgetId & id, const WidgetId & requestId, bool ro, const sva::PTransformd & pos)
-{
-  auto label = latestWidget_;
-  auto & w = get_widget<TransformInteractiveMarkerWidget>(id, requestId, impl_->int_server_, !ro, false, label);
-  w.update(pos);
-}
-
-void Panel::got_transform(const WidgetId & id, const WidgetId & requestId, bool ro, const sva::PTransformd & pos)
-{
-  auto label = latestWidget_;
-  auto & w = get_widget<TransformInteractiveMarkerWidget>(id, requestId, impl_->int_server_, !ro, !ro, label);
-  w.update(pos);
-}
-
-void Panel::got_xytheta(const WidgetId & id,
-                        const WidgetId & requestId,
-                        bool ro,
-                        const Eigen::Vector3d & vec,
-                        double altitude)
-{
-  auto label = latestWidget_;
-  auto & w = get_widget<XYThetaInteractiveMarkerWidget>(id, requestId, impl_->int_server_, sva::PTransformd::Identity(),
-                                                        !ro, !ro, label);
-  w.update(vec, altitude);
-}
-
-void Panel::got_trajectory(const WidgetId & id,
-                           const std::vector<Eigen::Vector3d> & points,
-                           const mc_rtc::gui::LineConfig & config)
-{
-  auto & w = get_widget<DisplayTrajectoryWidget>(id, impl_->marker_array_);
-  w.update(points, config);
-}
-
-void Panel::got_trajectory(const WidgetId & id,
-                           const std::vector<sva::PTransformd> & points,
-                           const mc_rtc::gui::LineConfig & config)
-{
-  auto & w = get_widget<DisplayTrajectoryWidget>(id, impl_->marker_array_);
-  w.update(points, config);
-}
-
-void Panel::got_trajectory(const WidgetId & id, const Eigen::Vector3d & point, const mc_rtc::gui::LineConfig & config)
-{
-  auto & w = get_widget<DisplayTrajectoryWidget>(id, impl_->marker_array_);
-  w.update(point, config);
-}
-
-void Panel::got_trajectory(const WidgetId & id, const sva::PTransformd & point, const mc_rtc::gui::LineConfig & config)
-{
-  auto & w = get_widget<DisplayTrajectoryWidget>(id, impl_->marker_array_);
-  w.update(point, config);
-}
-
-void Panel::got_polygon(const WidgetId & id,
-                        const std::vector<std::vector<Eigen::Vector3d>> & polygons,
-                        const mc_rtc::gui::LineConfig & c)
-{
-  auto & w = get_widget<PolygonMarkerWidget>(id, impl_->marker_array_);
-  w.update(polygons, c);
-}
-
-void Panel::got_polyhedron(const WidgetId & id,
-                           const std::vector<std::array<Eigen::Vector3d, 3>> & triangles,
-                           const std::vector<std::array<mc_rtc::gui::Color, 3>> & colors,
-                           const mc_rtc::gui::PolyhedronConfig & c)
-{
-  auto & w = get_widget<PolyhedronMarkerWidget>(id, impl_->marker_array_, c);
-  w.update(triangles, colors);
-}
-
-void Panel::got_force(const WidgetId & id,
-                      const WidgetId & requestId,
-                      const sva::ForceVecd & forcep,
-                      const sva::PTransformd & surface,
-                      const mc_rtc::gui::ForceConfig & forceConfig,
-                      bool ro)
-{
-  auto label = latestWidget_;
-  auto & w = get_widget<ForceInteractiveMarkerWidget>(id, requestId, impl_->int_server_, impl_->marker_array_, surface,
-                                                      forcep, forceConfig, ro, label);
-  w.update(surface, forcep, forceConfig);
-}
-
-void Panel::got_arrow(const WidgetId & id,
-                      const WidgetId & requestId,
-                      const Eigen::Vector3d & start,
-                      const Eigen::Vector3d & end,
-                      const mc_rtc::gui::ArrowConfig & config,
-                      bool ro)
-{
-  auto label = latestWidget_;
-  auto & w = get_widget<ArrowInteractiveMarkerWidget>(id, requestId, impl_->int_server_, impl_->marker_array_, start,
-                                                      end, config, !ro, !ro, label);
-  w.update(start, end, config);
-}
-
-void Panel::got_schema(const WidgetId & id, const std::string & schema)
-{
-  get_widget<SchemaWidget>(id, schema, data_);
-}
-
-void Panel::got_table_start(const WidgetId & id, const std::vector<std::string> & header)
-{
-  auto & w = get_widget<TableWidget>(id);
-  w.header(header);
-}
-
-void Panel::got_table_row(const WidgetId & id, const std::vector<std::string> & dataIn)
-{
-  auto & w = get_widget<TableWidget>(id);
-  w.row(dataIn);
-}
-
-void Panel::got_table_end(const WidgetId & id)
-{
-  auto & w = get_widget<TableWidget>(id);
-  w.finalize();
-}
-
-void Panel::got_robot(const WidgetId & id,
-                      const std::vector<std::string> & parameters,
-                      const std::vector<std::vector<double>> & /*q*/,
-                      const sva::PTransformd & /*posW*/)
-{
-  auto & w = get_widget<RobotModelDisplay>(id, displayContext(), displayGroup());
-  w.update(parameters[0]);
-}
-
-void Panel::got_visual(const WidgetId & id, const rbd::parsers::Visual & visual, const sva::PTransformd & pose)
-{
-  auto & w = get_widget<VisualWidget>(id, impl_->marker_array_);
-  w.update(visual, pose);
-}
-
-void Panel::got_form(const WidgetId & id)
-{
-  auto & form = get_widget<FormWidget>(id);
-  activeForm_ = form.container();
-  form.update();
-}
-
-void Panel::got_form_checkbox(const WidgetId & formId,
-                              const std::string & name,
-                              bool required,
-                              bool def,
-                              bool def_from_user)
-{
-  activeForm_->element<form::Checkbox>(name, required, def, def_from_user);
-}
-
-void Panel::got_form_integer_input(const WidgetId & formId,
-                                   const std::string & name,
-                                   bool required,
-                                   int def,
-                                   bool def_from_user)
-{
-  activeForm_->element<form::IntegerInput>(name, required, def, def_from_user);
-}
-
-void Panel::got_form_number_input(const WidgetId & formId,
-                                  const std::string & name,
-                                  bool required,
-                                  double def,
-                                  bool def_from_user)
-{
-  activeForm_->element<form::NumberInput>(name, required, def, def_from_user);
-}
-
-void Panel::got_form_string_input(const WidgetId & formId,
-                                  const std::string & name,
-                                  bool required,
-                                  const std::string & def,
-                                  bool def_from_user)
-{
-  activeForm_->element<form::StringInput>(name, required, def, def_from_user);
-}
-
-void Panel::got_form_array_input(const WidgetId & formId,
-                                 const std::string & name,
-                                 bool required,
-                                 const Eigen::VectorXd & def,
-                                 bool fixed_size,
-                                 bool def_from_user)
-{
-  activeForm_->element<form::NumberArrayInput>(name, required, def, fixed_size, def_from_user);
-}
-
-void Panel::got_form_combo_input(const WidgetId & formId,
-                                 const std::string & name,
-                                 bool required,
-                                 const std::vector<std::string> & values,
-                                 bool send_index,
-                                 int def)
-{
-  activeForm_->element<form::ComboInput>(name, required, values, send_index, def);
-}
-
-void Panel::got_form_data_combo_input(const WidgetId & formId,
-                                      const std::string & name,
-                                      bool required,
-                                      const std::vector<std::string> & ref,
-                                      bool send_index)
-{
-  activeForm_->element<form::DataComboInput>(name, required, data_, ref, send_index);
-}
-
-void Panel::got_form_point3d_input(const WidgetId & formId,
-                                   const std::string & name,
-                                   bool required,
-                                   const Eigen::Vector3d & default_,
-                                   bool default_from_user,
-                                   bool interactive)
-{
-  activeForm_->element<form::Point3DInput>(name, required, default_, default_from_user, interactive,
-                                           impl_->int_server_);
-}
-
-void Panel::got_form_rotation_input(const WidgetId & formId,
-                                    const std::string & name,
-                                    bool required,
-                                    const sva::PTransformd & default_,
-                                    bool default_from_user,
-                                    bool interactive)
-{
-  activeForm_->element<form::RotationInput>(name, required, default_, default_from_user, interactive,
-                                            impl_->int_server_);
-}
-
-void Panel::got_form_transform_input(const WidgetId & formId,
-                                     const std::string & name,
-                                     bool required,
-                                     const sva::PTransformd & default_,
-                                     bool default_from_user,
-                                     bool interactive)
-{
-  activeForm_->element<form::TransformInput>(name, required, default_, default_from_user, interactive,
-                                             impl_->int_server_);
-}
-
-void Panel::got_start_form_object_input(const std::string & name, bool required)
-{
-  activeForm_ = activeForm_->element<form::Object>(name, required, activeForm_)->container();
-}
-
-void Panel::got_end_form_object_input()
-{
-  activeForm_->update();
-  activeForm_ = activeForm_->parentForm();
-}
-
-void Panel::got_start_form_generic_array_input(const std::string & name,
-                                               bool required,
-                                               std::optional<std::vector<mc_rtc::Configuration>> data)
-{
-  auto array = activeForm_->element<form::GenericArray>(name, required, data, activeForm_);
-  activeForm_ = array->container();
-}
-
-void Panel::got_end_form_generic_array_input()
-{
-  activeForm_->update();
-  activeForm_ = activeForm_->parentForm();
-}
-
-void Panel::got_start_form_one_of_input(const std::string & name,
-                                        bool required,
-                                        const std::optional<std::pair<size_t, mc_rtc::Configuration>> & data)
-{
-  auto oneof = activeForm_->element<form::OneOf>(name, required, data, activeForm_);
-  oneof->update();
-  activeForm_ = oneof->container();
-}
-
-void Panel::got_end_form_one_of_input()
-{
-  activeForm_->update();
-  activeForm_ = activeForm_->parentForm();
-}
-
-void Panel::got_start_plot(uint64_t id, const std::string & title)
-{
-  got_category({}, "Plots");
-  auto & w = get_widget<PlotTabWidget>({{"Plots"}, ""});
-  w.start_plot(id, title);
-}
-
-void Panel::got_plot_setup_xaxis(uint64_t id, const std::string & legend, const mc_rtc::gui::plot::Range & range)
-{
-  auto & w = get_widget<PlotTabWidget>({{"Plots"}, ""});
-  w.plot_setup_xaxis(id, legend, range);
-}
-
-void Panel::got_plot_setup_yaxis_left(uint64_t id, const std::string & legend, const mc_rtc::gui::plot::Range & range)
-{
-  auto & w = get_widget<PlotTabWidget>({{"Plots"}, ""});
-  w.plot_setup_yaxis_left(id, legend, range);
-}
-
-void Panel::got_plot_setup_yaxis_right(uint64_t id, const std::string & legend, const mc_rtc::gui::plot::Range & range)
-{
-  auto & w = get_widget<PlotTabWidget>({{"Plots"}, ""});
-  w.plot_setup_yaxis_right(id, legend, range);
-}
-
-void Panel::got_plot_point(uint64_t id,
-                           uint64_t did,
-                           const std::string & legend,
-                           double x,
-                           double y,
-                           mc_rtc::gui::Color color,
-                           mc_rtc::gui::plot::Style style,
-                           mc_rtc::gui::plot::Side side)
-{
-  auto & w = get_widget<PlotTabWidget>({{"Plots"}, ""});
-  w.plot_point(id, did, legend, x, y, color, style, side);
-}
-
-void Panel::got_plot_polygon(uint64_t id,
-                             uint64_t did,
-                             const std::string & legend,
-                             const mc_rtc::gui::plot::PolygonDescription & polygon,
-                             mc_rtc::gui::plot::Side side)
-{
-  auto & w = get_widget<PlotTabWidget>({{"Plots"}, ""});
-  w.plot_polygon(id, did, legend, polygon, side);
-}
-
-void Panel::got_plot_polygons(uint64_t id,
-                              uint64_t did,
-                              const std::string & legend,
-                              const std::vector<mc_rtc::gui::plot::PolygonDescription> & polygons,
-                              mc_rtc::gui::plot::Side side)
-{
-  auto & w = get_widget<PlotTabWidget>({{"Plots"}, ""});
-  w.plot_polygons(id, did, legend, polygons, side);
-}
-
-void Panel::got_end_plot(uint64_t id)
-{
-  auto & w = get_widget<PlotTabWidget>({{"Plots"}, ""});
-  w.end_plot(id);
 }
 
 Panel::WidgetTree & Panel::get_category(const std::vector<std::string> & category)
